Seperated sudoku modules are merged into Sudoku.py.
